Fix a bug with calculating the DOES_TOWER_PROTECT_ENEMY environment variable
The DoesTowerProtectUnit function compares a distance from the bot
to the enemy tower and a distance from the target the tower. This
leads to cyclic stack when the bot moves front and back to the tower
and has a threshold distance value. The new version of the function
compares a distance from the target to the tower only.